refactor(updater): split boot_probation.rs into boot_probation/{schema,backup,detect,monitor,rollback,report}.rs
Behaviour-neutral move of the 1426-line boot_probation.rs into a
directory module. cfg-gated pairs (copy_symlink, helper_executable,
is_pid_alive) kept together with exact attributes; the external surface
is preserved via pub(crate) use re-exports in mod.rs. check green
(318 Rust + 1480 vitest tests).
